testes unitários do frindship, Controller e  Services
