+++ jaview_backend/src/__tests__/example.spec.ts
@@ -1,5 +1,5 @@
 describe('exemplo', () => {
     it('deve rodar o teste corretamente', () => {
-        expect(1+1).toBe(2);
+        expect(1 + 1).toBe(2);
     })
 })
